Random/snowflakegui/requirements.txt
gunicorn
psycopg2-binary
django-crispy-forms
Django==3.1.2
django-celery-results==1.2.1
django-cors-headers==3.5.0
snowflake-connector-python[pandas]
snowflake-sqlalchemy==1.2.3
celery==4.4.7
redis
ruamel.yaml
fastcore
nbdev
boto3
django-storages
